refactor(frontend): row-click → detail sheet + update AGENTS.md + CreateRevisionSheet
## Row-click opens detail (standardize across resource list pages)

Replaces the inconsistent mix of "Edit" (pencil) and "View" (eye) icon
buttons in the Operations column with a single pattern:

- The whole **row is clickable** to open the detail Sheet
- Row gets `cursor-pointer hover:bg-control-bg` as the affordance
- Operations column holds only **destructive/secondary actions** (Delete,
  Restore) — each with `e.stopPropagation()` so they don't also trigger
  row-open
- The "primary accent" color (`text-accent`) on the entity title is
  dropped in favor of `text-main` — the row hover state is the
  clickability signal, not a blue title

Whether the Sheet lets you edit or only view is a function of the user's
permissions (the form already disables inputs / hides Save when the user
lacks update permission). The entry point no longer promises "view" or
"edit" — both labels were half-truths on pages where a single row might
be editable for some users and read-only for others.

Applied to:
- `WorkloadIdentitiesPage` — removed Pencil edit button
- `UsersPage` — removed Eye/Pencil icon (whichever was shown per account
  type), kept Delete
- `GroupsPage` — removed Pencil edit button from GroupRow, kept Delete;
  the expand/collapse chevron button gets `e.stopPropagation()` so it
  toggles without also opening detail
- `ServiceAccountsPage` — removed Pencil edit button, kept Delete

## frontend/AGENTS.md: new Dialog vs Sheet guideline

Added a new section under `## shadcn Component Guidelines → Rules`
covering:

- When to use `<Sheet>` vs `<Dialog>` vs `<AlertDialog>` (creating/editing
  → Sheet; confirmation/prompt/read-only result → Dialog)
- Sheet width tiers (narrow / standard / wide) with concrete examples of
  when to pick each
- The outer-wrapper + inner-form + `key` pattern for edit sheets, with a
  code example, so future edit sheets populate reliably on every open
- The dirty-tracking rule (`allowConfirm = isFormValid && isDirty`) for
  edit sheets
- The fetch-full-entity-before-open rule to avoid the stub WI bug

## CreateRevisionDialog → CreateRevisionSheet

Complete the wrong-Dialog audit by migrating the last flagged offender.
Currently orphaned (not yet imported by DatabaseRevisionPanel which still
uses the Vue version), but in-place and ready for when that panel migrates.

Lint, type-check, 1031 unit tests all green.
