+++ frontend/src/react/pages/settings/ServiceAccountsPage.tsx
@@ -1,7 +1,7 @@
 import { create } from "@bufbuild/protobuf";
 import { FieldMaskSchema } from "@bufbuild/protobuf/wkt";
-import { Copy, KeyRound, Pencil, Plus, Trash2, Undo2 } from "lucide-react";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { Copy, KeyRound, Plus, Trash2, Undo2 } from "lucide-react";
+import { useCallback, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { PermissionGuard } from "@/react/components/PermissionGuard";
 import { RoleSelect } from "@/react/components/RoleSelect";
@@ -20,6 +20,7 @@ import {
 import { Tooltip } from "@/react/components/ui/tooltip";
 import { PagedTableFooter, usePagedData } from "@/react/hooks/usePagedData";
 import { useVueState } from "@/react/hooks/useVueState";
+import { cn } from "@/react/lib/utils";
 import {
   ensureServiceAccountFullName,
   pushNotification,
@@ -169,11 +170,23 @@ function ServiceAccountTable({
           ) : (
             users.map((user, i) => {
               const isDeleted = user.state === State.DELETED;
+              const canOpenDetail =
+                !!onUserSelected &&
+                (project
+                  ? hasProjectPermissionV2(project, "bb.serviceAccounts.get")
+                  : hasWorkspacePermissionV2("bb.serviceAccounts.get"));
 
               return (
                 <tr
                   key={user.name}
-                  className={`border-b last:border-b-0 ${i % 2 === 1 ? "bg-control-bg" : ""}`}
+                  className={cn(
+                    "border-b last:border-b-0",
+                    i % 2 === 1 && "bg-control-bg/50",
+                    canOpenDetail && "cursor-pointer hover:bg-control-bg"
+                  )}
+                  onClick={
+                    canOpenDetail ? () => onUserSelected(user) : undefined
+                  }
                 >
                   {/* Account column */}
                   <td className="px-4 py-2">
@@ -185,7 +198,7 @@ function ServiceAccountTable({
                             className={
                               isDeleted
                                 ? "line-through text-control-light"
-                                : "font-medium text-accent"
+                                : "font-medium text-main"
                             }
                           >
                             {user.title}
@@ -260,55 +273,37 @@ function ServiceAccountTable({
                     </div>
                   </td>
 
-                  {/* Operations column */}
+                  {/* Operations column — destructive/secondary actions only.
+                      The row itself is clickable to open the detail sheet. */}
                   <td className="px-4 py-2">
                     <div className="flex justify-end gap-x-1">
-                      {!isDeleted && (
-                        <>
-                          {(project
-                            ? hasProjectPermissionV2(
-                                project,
-                                "bb.serviceAccounts.delete"
-                              )
-                            : hasWorkspacePermissionV2(
-                                "bb.serviceAccounts.delete"
-                              )) && (
-                            <Tooltip
-                              content={t(
-                                "settings.members.action.deactivate-confirm-title"
-                              )}
+                      {!isDeleted &&
+                        (project
+                          ? hasProjectPermissionV2(
+                              project,
+                              "bb.serviceAccounts.delete"
+                            )
+                          : hasWorkspacePermissionV2(
+                              "bb.serviceAccounts.delete"
+                            )) && (
+                          <Tooltip
+                            content={t(
+                              "settings.members.action.deactivate-confirm-title"
+                            )}
+                          >
+                            <Button
+                              variant="ghost"
+                              size="icon"
+                              className="h-7 w-7 text-error hover:text-error"
+                              onClick={(e) => {
+                                e.stopPropagation();
+                                handleDeactivate(user);
+                              }}
                             >
-                              <Button
-                                variant="ghost"
-                                size="icon"
-                                className="h-7 w-7 text-error hover:text-error"
-                                onClick={() => handleDeactivate(user)}
-                              >
-                                <Trash2 className="h-4 w-4" />
-                              </Button>
-                            </Tooltip>
-                          )}
-                          {(project
-                            ? hasProjectPermissionV2(
-                                project,
-                                "bb.serviceAccounts.get"
-                              )
-                            : hasWorkspacePermissionV2(
-                                "bb.serviceAccounts.get"
-                              )) && (
-                            <Tooltip content={t("common.edit")}>
-                              <Button
-                                variant="ghost"
-                                size="icon"
-                                className="h-7 w-7"
-                                onClick={() => onUserSelected?.(user)}
-                              >
-                                <Pencil className="h-4 w-4" />
-                              </Button>
-                            </Tooltip>
-                          )}
-                        </>
-                      )}
+                              <Trash2 className="h-4 w-4" />
+                            </Button>
+                          </Tooltip>
+                        )}
                       {isDeleted &&
                         (project
                           ? hasProjectPermissionV2(
@@ -327,7 +322,10 @@ function ServiceAccountTable({
                               variant="ghost"
                               size="icon"
                               className="h-7 w-7"
-                              onClick={() => handleRestore(user)}
+                              onClick={(e) => {
+                                e.stopPropagation();
+                                handleRestore(user);
+                              }}
                             >
                               <Undo2 className="h-4 w-4" />
                             </Button>
@@ -364,10 +362,7 @@ function CreateServiceAccountSheet(props: CreateServiceAccountSheetProps) {
     <Sheet open={open} onOpenChange={(next) => !next && onClose()}>
       <SheetContent width="standard">
         {open && (
-          <ServiceAccountForm
-            key={serviceAccount?.name ?? "new"}
-            {...props}
-          />
+          <ServiceAccountForm key={serviceAccount?.name ?? "new"} {...props} />
         )}
       </SheetContent>
     </Sheet>
@@ -539,79 +534,79 @@ function ServiceAccountForm({
       </SheetHeader>
 
       <SheetBody>
-          <div className="flex flex-col gap-y-6">
-            {/* Name */}
-            <div className="flex flex-col gap-y-2">
-              <label className="block text-sm font-medium text-control">
-                {t("common.name")}
-              </label>
-              <Input
-                autoComplete="off"
-                value={title}
-                onChange={(e) => setTitle(e.target.value)}
-                placeholder="Foo"
-                maxLength={200}
-              />
-            </div>
-
-            {/* Email */}
-            <div className="flex flex-col gap-y-2">
-              <label className="block text-sm font-medium text-control">
-                {t("common.email")}
-                <span className="ml-0.5 text-error">*</span>
-              </label>
-              {isEditMode ? (
-                <Input value={serviceAccount?.email ?? ""} disabled />
-              ) : (
-                <div className="px-1 flex items-center border border-control-border rounded-xs overflow-hidden focus-within:border-accent">
-                  <input
-                    type="text"
-                    autoComplete="off"
-                    value={emailPrefix}
-                    onChange={(e) => setEmailPrefix(e.target.value)}
-                    className="flex-1 h-9 px-3 py-1 text-sm bg-transparent border-none outline-none ring-0 shadow-none focus:border-none focus:outline-none focus:ring-0 focus:shadow-none"
-                  />
-                  <span className="px-2 py-1 text-sm text-control-light bg-control-bg whitespace-nowrap">
-                    @{emailSuffix}
-                  </span>
-                </div>
-              )}
-            </div>
-
-            {/* Roles (create mode only) */}
-            {!isEditMode &&
-              (projectEntity
-                ? hasProjectPermissionV2(
-                    projectEntity,
-                    "bb.projects.setIamPolicy"
-                  )
-                : hasWorkspacePermissionV2("bb.workspaces.setIamPolicy")) && (
-                <div className="flex flex-col gap-y-2">
-                  <label className="block text-sm font-medium text-control">
-                    {t("settings.members.table.roles")}
-                  </label>
-                  <RoleSelect
-                    value={roles}
-                    onChange={setRoles}
-                    disabled={false}
-                    scope={project ? "project" : undefined}
-                  />
-                </div>
-              )}
+        <div className="flex flex-col gap-y-6">
+          {/* Name */}
+          <div className="flex flex-col gap-y-2">
+            <label className="block text-sm font-medium text-control">
+              {t("common.name")}
+            </label>
+            <Input
+              autoComplete="off"
+              value={title}
+              onChange={(e) => setTitle(e.target.value)}
+              placeholder="Foo"
+              maxLength={200}
+            />
           </div>
-        </SheetBody>
 
-        <SheetFooter>
-          <Button variant="outline" onClick={onClose}>
-            {t("common.cancel")}
-          </Button>
-          <Button
-            disabled={!allowConfirm || !hasPermission || isRequesting}
-            onClick={handleSubmit}
-          >
-            {isEditMode ? t("common.update") : t("common.create")}
-          </Button>
-        </SheetFooter>
+          {/* Email */}
+          <div className="flex flex-col gap-y-2">
+            <label className="block text-sm font-medium text-control">
+              {t("common.email")}
+              <span className="ml-0.5 text-error">*</span>
+            </label>
+            {isEditMode ? (
+              <Input value={serviceAccount?.email ?? ""} disabled />
+            ) : (
+              <div className="px-1 flex items-center border border-control-border rounded-xs overflow-hidden focus-within:border-accent">
+                <input
+                  type="text"
+                  autoComplete="off"
+                  value={emailPrefix}
+                  onChange={(e) => setEmailPrefix(e.target.value)}
+                  className="flex-1 h-9 px-3 py-1 text-sm bg-transparent border-none outline-none ring-0 shadow-none focus:border-none focus:outline-none focus:ring-0 focus:shadow-none"
+                />
+                <span className="px-2 py-1 text-sm text-control-light bg-control-bg whitespace-nowrap">
+                  @{emailSuffix}
+                </span>
+              </div>
+            )}
+          </div>
+
+          {/* Roles (create mode only) */}
+          {!isEditMode &&
+            (projectEntity
+              ? hasProjectPermissionV2(
+                  projectEntity,
+                  "bb.projects.setIamPolicy"
+                )
+              : hasWorkspacePermissionV2("bb.workspaces.setIamPolicy")) && (
+              <div className="flex flex-col gap-y-2">
+                <label className="block text-sm font-medium text-control">
+                  {t("settings.members.table.roles")}
+                </label>
+                <RoleSelect
+                  value={roles}
+                  onChange={setRoles}
+                  disabled={false}
+                  scope={project ? "project" : undefined}
+                />
+              </div>
+            )}
+        </div>
+      </SheetBody>
+
+      <SheetFooter>
+        <Button variant="outline" onClick={onClose}>
+          {t("common.cancel")}
+        </Button>
+        <Button
+          disabled={!allowConfirm || !hasPermission || isRequesting}
+          onClick={handleSubmit}
+        >
+          {isEditMode ? t("common.update") : t("common.create")}
+        </Button>
+      </SheetFooter>
     </>
   );
 }
